pyproject.toml
[project]
name = "res-forecasting"
version = "0.1.0"
description = ""
authors = [
    {name = "Dimosthenis Tsagkrasoulis",email = "dtsagkrasoulis@i4.energy"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "tinydb (>=4.8.2,<5.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.setuptools.package-data]
"res_forecasting.data" = ["weather_data.json"]
[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
mypy = "^1.16.1"
black = "^25.1.0"

